Introduce sine wave crest factor to RMS calculation.
The RMS of a full sine wave is -3.01dBFS, not 0dBFS. Since music is more or
less a sine wave, it makes sense to use that as our reference point.

A sine wave's crest factor is the square root of 2, or 3.01dB. There are
a few places said factor could be introduced:

1. In the RMS calculation, by multiplying it by the crest factor.
2. In the RMS calculation, by introducing a factor of 2.0 inside the
   sqrt() call. (Essentially the same as option 1.)
3. After the final average has been calulated and converted to dB, by
   adding 3.01dB.

I have arbitrarily chosen the second option.

This change puts us *very close* to the official DR meter—meaning, within
a few tenths of the official values (based on information gleaned from logs
posted online; i have not yet been able to compare apples to apples).

https://en.wikipedia.org/wiki/Crest_factor