Support other sample sizes.
We now support every sample size that ffmpeg does. And that pretty much
rounds things out, in terms of format support. We should now be able to
handle any type of audio that ffmpeg can.

There is a slight performance penalty, due to introducing a switch
statement into the scanning loop.